refactor: move resource-extensions.ts out of commands/
