+++ .npmrc
@@ -1,0 +1,1 @@
+script-shell=C:\Program Files\Git\bin\bash.exe
