+++ src/deploy-helpers.ts
@@ -18,11 +18,11 @@ import { basename, dirname, extname, join, relative, resolve } from "node:path";
 import { TenantId } from "@camunda8/orchestration-cluster-api";
 import type { Ignore } from "ignore";
 import { createClient } from "./client.ts";
-import { DEPLOYABLE_EXTENSIONS } from "./commands/resource-extensions.ts";
 import { resolveTenantId } from "./config.ts";
 import { normalizeToError, SilentError } from "./errors.ts";
 import { isIgnored, loadIgnoreRules, resolveIgnoreBaseDir } from "./ignore.ts";
 import { getLogger, isRecord } from "./logger.ts";
+import { DEPLOYABLE_EXTENSIONS } from "./resource-extensions.ts";
 import { c8ctl } from "./runtime.ts";
 
 const PROCESS_APPLICATION_FILE = ".process-application";
