+++ src/resource-extensions.ts
